+++ lib/features/repository/presentation/ui/repository_screen.dart
@@ -70,6 +70,40 @@ class _RepositoryScreenState extends State<RepositoryScreen> {
             listenWhen: (previous, current) => previous.status != current.status,
             listener: (context, state) async {
               switch (state.status) {
+                case WorkingDirectoryBlocStatus.askForDiscardChanges:
+                  await showDialog<bool>(
+                    barrierDismissible: false,
+                    context: context,
+                    builder: (_) {
+                      // TODO : Create DiscardAllChangesDialog
+                      return AlertDialog(
+                        title: const Text("Discard all changes"),
+                        content: const Text(
+                          "This will permanently discard all local changes. This action cannot be undone.",
+                        ),
+                        actions: [
+                          TextButton(
+                            onPressed: () {
+                              _workingDirectoryBloc.add(
+                                UpdateWorkingDirectoryStatus(
+                                  status: WorkingDirectoryBlocStatus.initial,
+                                ),
+                              );
+                              Navigator.pop(context);
+                            },
+                            child: const Text("Cancel"),
+                          ),
+                          ElevatedButton(
+                            onPressed: () {
+                              Navigator.pop(context);
+                              _workingDirectoryBloc.add(DiscardAllChanges());
+                            },
+                            child: const Text("Discard"),
+                          ),
+                        ],
+                      );
+                    },
+                  );
                 case WorkingDirectoryBlocStatus.commitsPushed:
                   context.read<CommitHistoryBloc>().add(LoadCommitHistory());
                   _workingDirectoryBloc.add(
@@ -250,6 +284,7 @@ class _RepositoryScreenState extends State<RepositoryScreen> {
                         _workingDirectoryBloc.add(PushCommits());
                       },
                       isLoading: wdState.status == WorkingDirectoryBlocStatus.loading,
+                      hasUpstream: wdState.hasUpstream,
                     );
                   },
                 ),
@@ -260,9 +295,6 @@ class _RepositoryScreenState extends State<RepositoryScreen> {
                         children: [
                           RepositorySidebar(
                             files: workingDirectoryState.files,
-                            onFileSelected: (file) {
-                              _filesDifferencesBloc.add(LoadFileDiff(file: file));
-                            },
                             hasStagedFiles: workingDirectoryState.files.any((file) => file.staged),
                             onCommitPressed: ({required description, required summary}) {
                               _workingDirectoryBloc.add(AddCommit(summary: summary, description: description));