+++ lib/shared/core/constants/git_commands.dart
@@ -14,4 +14,8 @@ class GitCommands {
   static const List<String> gitPush = ["push"];
   static const List<String> remoteVerbose = ["remote", "-v"];
   static const List<String> remoteGetOrigin = ["remote", "get-url", "origin"];
+  static const List<String> getUpstreamState = ["rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{u}"];
+  static const List<String> publishBranch = ["push", "-u", "origin", "HEAD"];
+  static const List<String> restoreTrackedFiles = ["restore", "."];
+  static const List<String> removeUntrackedFiles = ["clean", "-fd"];
 }